Adds AIComponent to just hold the button for now
