+++ Danki2/Assets/Scripts/AI/AIComponent.cs
@@ -1,0 +1,5 @@
+﻿using UnityEngine;
+
+public class AIComponent : MonoBehaviour
+{
+}
