+++ Danki2/Assets/Scripts/AI/AIComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19514c5d500c2bd459b1f556adf129a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
