+++ Danki2/Assets/Scripts/AI/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b502505bdf61bba479940b10de40454a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
