+++ Danki2/Assets/Scripts/AI/Editor/AIComponentEditor.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using UnityEditor;
+using UnityEngine;
+
+[CustomEditor(typeof(AIComponent))]
+public class AIComponentEditor : Editor
+{
+    public override void OnInspectorGUI()
+    {
+        base.OnInspectorGUI();
+
+        if (GUILayout.Button("Log Attribute Values"))
+        {
+            BehaviourScanner.Scan();
+            LogBehaviourData();
+        }
+    }
+
+    private void LogBehaviourData()
+    {
+        foreach (AIAction action in Enum.GetValues(typeof(AIAction)))
+        {
+            var data = BehaviourScanner.GetDataByAction(action);
+            string message = action.ToString() + " behaviours found: " + data.Count;
+            if (data.Count > 0) message += ": ";
+            foreach (var d in data)
+            {
+                message += d.DisplayValue + ", ";
+            }
+            message.Remove(message.Length - 2);
+
+            Debug.Log(message);
+        }
+
+    }
+}
